== Office ==  * [CHANGED] Rewrite of M_Office_Actions.  * [NEW] Chainability with office actions  * [FIXED] Batch actions with parameters form can now be multi-pages  * [FIXED] Delete action can now be applied on whole resultset
== Plugins ==
 * [CHANGED] better event handling.
 * [FIXED] plugin tag working better

git-svn-id: https://m4php5.googlecode.com/svn/trunk@72 4cff5740-f916-11dd-b507-8b4ecd5210be

+++ Plugins/tag/DB/tag.php
@@ -12,24 +12,32 @@
 /**
  * Manages tags for any database record.
  * Logs tag add/remove history
- * This plugin needs 3 tables (@see install.php)
+ * This plugin needs 3 tables (@see Plugins/tag/commandes/install.php)
  */
 
 class DB_DataObject_Plugin_Tag extends M_Plugin {
 
   public function getEvents()
   {
-    return array('addtag','removetag','removetags','getbytags','postdelete');
+    return array('addtag','removetag','removetags','getbytags','postdelete','hastag','gettaglasthistory');
   }
   
   /**
    * adds a tag to a record
-   * @param DataObject_Tag tag to add
+   * @param mixed : DataObject_Tag tag to add or string.
    * @param DB_DataObject database record to tag
    */
 
-  public function addTag(DataObject_Tag $tag, DB_DataObject $obj)
+  public function addTag($tag, DB_DataObject $obj)
   {
+    if(!$existingtag = $this->_getTagFromTag($tag)) {
+      $newtag = DB_DataObject::factory('tag');
+      $newtag->strip = $tag;
+      $newtag->insert();
+      $tag = $newtag;
+    } else {
+      $tag = $existingtag;
+    }
     $dbo = DB_DataObject::factory('tag_record');
     $dbo->tag_id = $tag->id;
     $dbo->record_id = $obj->pk();
@@ -37,7 +45,7 @@ class DB_DataObject_Plugin_Tag extends M_Plugin {
     if(!$dbo->find(true)) {
       $dbo->insert();
     }
-    return;
+    return $this->returnStatus($obj);
   }
 
   /**
@@ -49,13 +57,15 @@ class DB_DataObject_Plugin_Tag extends M_Plugin {
 
   public function removeTag(DataObject_Tag $tag, DB_DataObject $obj)
   {
+    
+    if(!$tag = $this->_getTagFromTag($tag)) {return $this->returnStatus($obj);} 
     $dbo = DB_DataObject::factory('tag_record');
     $dbo->setTag($tag);
     $dbo->setRecord($obj);
     if($dbo->find(true)) {
       $dbo->delete();
     }
-    return;
+    return $this->returnStatus($obj);
   }
 
 
@@ -72,7 +82,7 @@ class DB_DataObject_Plugin_Tag extends M_Plugin {
     while($dbo->fetch()) {
       $dbo->delete();
     }
-    return;
+    return $this->returnStatus($obj);
   }
 
 
@@ -85,6 +95,7 @@ class DB_DataObject_Plugin_Tag extends M_Plugin {
   public function getByTags($tags, DB_DataObject $obj)
   {
     foreach($tags as $atag) {
+    if(!$atag = $this->_getTagFromTag($atag)) {continue;}
       $tagsid[] = $atag->pk();
     }
     $dbo = DB_DataObject::factory('tag_record');
@@ -94,21 +105,58 @@ class DB_DataObject_Plugin_Tag extends M_Plugin {
     $obj->joinAdd($dbo);
     return;
   }
-  public function getTagged(DB_DataObject $tag)
+  public function getTagged($tag)
   {
+    if(!$tag = $this->_getTagFromTag($tag)) {return $this->returnStatus(false);}
     $dbo = DB_DataObject::factory('tag_record');
     $dbo->setTag($tag);
     $dbo->find();
     return $dbo;
   }
-  public static function countTagged(DB_DataObject $tag)
+  public function hastag($tag,$obj)
   {
+    if(!$tag = $this->_getTagFromTag($tag)) {return $this->returnStatus(false);}
+    $dbo = DB_DataObject::factory('tag_record');
+    $dbo->tag_id = $tag->id;
+    $dbo->setRecord($obj);
+    $res = $dbo->count();
+    if($res<1) {
+        return $this->returnStatus(false);
+    }
+    return $this->returnStatus(true);
+  }
+  public function getTagLastHistory($tag,$direction,DB_DataObject $obj)
+  {
+    if(!$tag = $this->_getTagFromTag($tag)) {return $this->returnStatus(false);}
+    $h = DB_DataObject::factory('tag_history');
+    $h->tag_id = $tag->id;
+    $h->record_id = $obj->pk();
+    $h->tagged_table = $obj->tableName();
+    $h->direction=$direction;
+    $h->orderBy('date DESC');
+    $h->find(true);
+    return $this->returnStatus($h);
+  }
+  public static function countTagged($tag)
+  {
+        if(!$tag = $this->_getTagFromTag($tag)) {return $this->returnStatus(0);}
     $dbo = DB_DataObject::factory('tag_record');
     $dbo->setTag($tag);
-    return $dbo->count();
+    return $this->returnStatus($dbo->count());
   }
   public function postdelete($obj)
   {
     $this->removeTags($obj);
+  }
+  protected function _getTagFromTag($tag) {
+    if(!is_a($tag,'DataObjects_Tag')) {
+      $t = DB_DataObject::factory('tag');
+      $t->strip = $tag;
+      if(!$t->find(true)) {
+        return false;
+      }
+      $tag = $t;
+    }
+    return $tag;
   }
 }
